requirements.txt
uvicorn==0.19.0
fastapi==0.85.1
pillow==9.2.0
matplotlib==3.6.1
quickdraw==0.2.0
numpy==1.23.4
tensorboard==2.10.1
opencv-python==4.6.0.66